Add main menu scene and UI system improvements
Introduces a main menu scene with background and play button entities, including new components and tags for UI. Refactors rendering systems to support UI scaling and text spacing, updates scene management and ambient sound logic, and improves vector and bounds utilities. Also updates CMake configuration and main loop structure for better modularity.

+++ include/systems/core/rendering/sprite_render_system.h
@@ -15,7 +15,7 @@ public:
 
 	static void DrawUI(
 		entt::registry& registry,
-		Nc::Vector2i windowSize
+		Nc::Vector2f windowSize
 	);
 
 };
